+++ backends/carp_backend/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+
+* require `carp_mobile_sensing` ^3.0.0
+
 ## 2.1.1
 
 - upgrade to `research_package` ^3.0.0